Changing directory structure and adding Gulp.
This got a little out of control. Basically, I wanted to change the directory
structure so that it was clear where the `client_app` file would live in
relation to other server-side code.

I also backed off of using Webpack for everything. Instead, I opted to use Gulp
for the scss and Browsersync functionality. I may be missing out on some
webpack-dev-server features but that can be corrected later. If you're
interested in why I chose to use Gulp in addition to webpack, check out the
README.